==> 03_1_Tratamento_Basico.ipynb <==
## Script responsável pelo tratamento básico

- Este código pressupõe que a existência do arquivo **.\saida\02_DouSecao0<No Seção>_portarias_mgi_tratado.parquet** 
- Caso não exista, ele pode ser gerado por meio do notebook **02_Tratamento_AnaliseDescritiva.ipynb**
- Processamento deste script:
  - Faz a leitura do arquivo **./saida/01_DOUSecao02_completo.parquet**
  - **Tratamento 01**
  - Coloca em minúsculo
  - Remove espaços extras
  - Remove acentos
  - Remove URLs
  - Remove valores monetários
  - Remove stopover
  - Armazena na **coluna TextoTratado**
  - Cria a **coluna TextoTratado_TotalPalavras** com o total do palavras do texto tratado
  
---


## Bibliotecas

# Biblioteca para manipulação e análise de dados
import pandas as pd

# Bibliotecas para processamento de texto
import re  # Para operações com expressões regulares
import unicodedata  # Para normalização de texto

# Biblioteca NLTK para processamento de linguagem natural
import nltk
from nltk.corpus import stopwords


douItem = 2
secao = f'Secao0{douItem}'

# Lê o arquivo Parquet
df_portarias_mgi = pd.read_parquet(f'./saida/02_DOU{secao}_portarias_mgi_tratado.parquet', engine='pyarrow')  # ou engine='fastparquet'

df_portarias_mgi.shape


# Baixar stopwords caso ainda não tenha feito
# nltk.download('stopwords')

# Obter stopwords em português
stop_words = set(stopwords.words('portuguese'))

# Função para remover acentos
def remover_acentos(text):
    """
    Remove acentos de um texto.
    
    Args:
        text (str): Texto com possíveis acentos.
    
    Returns:
        str: Texto sem acentos.
    """
    return ''.join(
        c for c in unicodedata.normalize('NFD', text) 
        if unicodedata.category(c) != 'Mn'
    )

# Função para processar o texto
def process_text(text):
    """
    Processa o texto, removendo URLs, acentos, valores monetários, 
    stopwords e o texto 'portaria de pessoal'.
    
    Args:
        text (str): Texto original.
    
    Returns:
        str: Texto processado.
    """
    text = text.lower()  # Colocar o texto em minúsculas
    text = " ".join(text.split())  # Remove espaços extras
    text = re.sub(r"http\S+|www.\S+", "", text)  # Remove URLs
    # Remover valores monetários
    text = re.sub(r"r\$\s?\d{1,3}(\.\d{3})*,\d{2}", "", text, flags=re.IGNORECASE)
    # Remover o texto 'portaria de pessoal'
    text = re.sub(r"portaria de pessoal", "", text, flags=re.IGNORECASE)
    text = remover_acentos(text)  # Remove os acentos
    text = " ".join([word for word in text.split() if word not in stop_words])  # Remove stopwords
    return text

# Aplicar a função no DataFrame
df_portarias_mgi['TextoTratado'] = df_portarias_mgi['Texto'].apply(process_text)

# Exibir o shape do DataFrame
df_portarias_mgi.shape


# Criar a nova coluna com o total de palavras
df_portarias_mgi['TextoTratado_TotalPalavras'] = df_portarias_mgi['TextoTratado'].apply(
    lambda texto: len(str(texto).split())
)

df_portarias_mgi.shape

df_portarias_mgi.head()

# Salva saída parcial o arquivo completo com novas colunas
df_portarias_mgi.to_parquet(f'./saida/03_1_DOU{secao}_portarias_mgi_tratado_basico.parquet', engine='pyarrow', index=False)
